Create README.md
This is an active World Map made with Python, which provides the information about the Population as well as key aspects of the country such as Latitudes and Longitudes. The borders are added by reading and uploading data from a JSON file. The countries are differentiated with different colors based on their Population(millions) according to the Year 2005. I also added some Markers of active Volcanoes in the region of United States Of America with their important aspects such as Elevation Level, Name, Status, Latitudes and Longitudes.